Software_Engineering_for_Big_Data/ESBD3_distributed_systems/1.crop_recommendation/crop_recommendation/requirements.txt
numpy
pandas
scikit-learn
joblib
pyzmq
minio
